// vendor/raw.githubusercontent.com/#mikaelporttila_0c368/rss/refs/tags/1.0.3/src/types/fields/rss2_fields.ts
export enum Rss2Fields {
  Author = "author",
  Category = "category",
  Channel = "channel",
  Cloud = "cloud",
  Copyright = "copyright",
  Description = "description",
  Docs = "docs",
  Enclosure = "enclosure",
  Generator = "generator",
  Guid = "guid",
  Height = "height",
  Image = "image",
  isPermaLink = "ispermalink",
  Item = "item",
  Language = "language",
  LastBuildDate = "lastbuilddate",
  Length = "length",
  Link = "link",
  ManagingEditor = "managingeditor",
  PubDate = "pubdate",
  SkipDays = "skipdays",
  SkipHours = "skiphours",
  Source = "source",
  TextInput = "textinput",
  Title = "title",
  Ttl = "ttl",
  Url = "url",
  WebMaster = "webmaster",
  Width = "width",
  Hour = "hour",
  Day = "day",
  ContentEncoded = "content:encoded",
  FeedburnerOrigLink = "feedburner:origlink",
}